requirements/test.txt
-r base.txt

djangorestframework
